Update Systemd-Unit-File.md
Added ExecStop commands as they were missing.

Changed default directory to '/opt/uptime-kuma' as it that is where the installer defaulted to in install.sh and will save the user having to manually make that change.

Removed the 'uptime' user requirement as that installation method isn't viable. This is due to, as I understand, the necessity of root privileges being required in order to use socket capabilities in the linux kernel via CAP_NET_RAW.

I have followed advice on setting the setuid bit for the process, that wasn't a good work around as it was a heavy security risk. I also tried granting the process and related server.js file the CAP_NET_RAW capability via the following command: 'setcap cap_net_raw+ep <file_name>' to no avail.

See the following for more info on that subject:

    https://stackoverflow.com/questions/20763039/creating-raw-socket-in-python-without-root-privileges.
    https://serverfault.com/questions/390153/which-permission-is-needed-to-open-a-tcp-server-socket-on-linux
    https://stackoverflow.com/questions/9772068/raw-socket-access-as-normal-user-on-linux-2-4